backend/requirements.txt
aenum==3.1.12
affine==2.3.1
aiofiles==0.5.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.7.7
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asf-search==6.0.0
asgiref==3.5.2
async-timeout==4.0.3
asynctest==0.13.0
attrs==22.1.0
awscli==1.31.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
botocore==1.33.2
bqplot==0.12.42
branca==0.7.0
cachetools==4.2.4
Cartopy==0.20.3
certifi==2024.2.2
cffi==1.15.0
cfgrib==0.9.10.4
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
cmocean==2.0
colorama==0.4.4
colour==0.1.5
comm==0.1.4
coverage==5.5
cryptography==35.0.0
cycler==0.11.0
dask==2022.2.0
dataclasses-json==0.5.7
dateparser==1.1.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
docutils==0.16
earthengine-api==0.1.384
eccodes==1.2.0
eccodes-python==0.9.9
ecdsa==0.17.0
ecmwflibs==0.5.3
ee-extra==0.0.15
eerepr==0.0.4
entrypoints==0.4
eyed3==0.9.7
fastapi==0.68.1
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
filetype==1.2.0
findlibs==0.0.5
Fiona==1.8.22
Flask==2.2.5
folium==0.15.1
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
GDAL==3.6.2
gdown==4.7.1
geeadd==1.0.1
geedim==1.7.2
geemap==0.20.4
geesarfetcher==0.3.8
geetools==0.6.14
geocoder==1.38.1
geojson==2.5.0
geopandas==0.10.2
geos==0.2.3
google-api-core==2.10.2
google-api-python-client==2.111.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.22.0
idna==3.4
imageio==2.26.0
importlib-metadata==4.13.0
importlib-resources==5.10.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.16.2
ipyleaflet==0.18.1
ipython==7.34.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.8.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-server==1.23.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.7
kiwisolver==1.4.4
laspy==2.5.0
locket==1.0.0
logzero==1.7.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
metview==1.14.0
mgrs==1.4.5
mistune==2.0.5
motor==2.4.0
mpmath==1.2.1
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.72
OSR==0.0.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
passlib==1.7.4
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2022.11.5
pkgutil_resolve_name==1.3.10
platformdirs==2.5.3
plotly==5.18.0
progressbar==2.5
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
PyCRS==1.0.2
pydantic==1.9.0
PyGeodesy==23.7.12
pyglet==1.1.4
Pygments==2.14.0
pygrib==2.1.4
pymongo==3.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
pyproj==3.2.1
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
python-box==7.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
python-snappy==0.6.1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyulog==1.1.0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
rasterio==1.2.10
ratelim==0.1.6
regex==2022.3.2
remotezip==0.10.0
requests==2.28.1
requests-oauthlib==1.3.1
rioxarray==0.9.1
rsa==4.7.2
s3transfer==0.8.1
sankee==0.2.4
scikit-fuzzy==0.4.2
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.7.3
scooby==0.7.3
Send2Trash==1.8.0
sentinelhub==3.8.4
shapely==2.0.1
shared-memory38==0.1.2
six==1.16.0
skimpy==0.8.4
sniffio==1.3.0
snuggs==1.4.7
sobol-seq==0.2.0
soupsieve==2.3.2.post1
spectral==0.23.1
SQLAlchemy==1.4.25
starlette==0.14.2
statsmodels==0.13.5
surfer==0.0.1
sympy==1.10.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
typing-inspect==0.9.0
typing_extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
uavsar-pytools==0.7.1
uritemplate==4.1.1
urllib3==1.26.12
utm==0.7.0
uvicorn==0.13.4
virtualenv==20.16.6
virtualenv-clone==0.5.7
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==3.6.6
xarray==0.20.2
xgboost==1.6.2
xyzservices==2023.2.0
yarl==1.9.4
zipp==3.10.0
